+++ lib/src/devices/soundcore/a3062/packets/outbound.rs
@@ -26,6 +26,10 @@ pub fn set_equalizer<
     equalizer_configuration: &CommonEqualizerConfiguration<CHANNELS, BANDS>,
     hear_id: &CustomHearId<HEAR_ID_CHANNELS, HEAR_ID_BANDS>,
 ) -> packet::Outbound {
+    // this should check if hear id is enabled, but that's not really necessary since set_equalizer
+    // will never be called with hear id enabled.
+    let active_volume_adjustments = equalizer_configuration.volume_adjustments();
+
     let body = equalizer_configuration
         .preset_id()
         .to_le_bytes()
@@ -43,12 +47,7 @@ pub fn set_equalizer<
         .chain(hear_id.time.to_be_bytes())
         .chain(iter::once(hear_id.hear_id_type as u8))
         .chain(hear_id.custom_volume_adjustment_bytes())
-        .chain(
-            equalizer_configuration
-                .volume_adjustments()
-                .iter()
-                .flat_map(|v| v.apply_drc().bytes()),
-        )
+        .chain(active_volume_adjustments.iter().flat_map(|v| v.bytes()))
         .chain(iter::once(0)) // unknown
         .collect();
     packet::Outbound::new(packet::Command([0x03, 0x87]), body)